requirements.txt
celery==4.2.1
Django>=2.2,<3
djangorestframework==3.12.4
django-pg-bulk-update==3.4.3
django-pg-returning==1.3.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
requests
sentry-sdk==1.3.1
statsd
typing
